+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/shared/attachment.py
@@ -11,7 +11,6 @@ from typing import Optional
 
 @dataclasses.dataclass
 class Attachment:
-    r"""OK"""
     company_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('companyId'), 'exclude': lambda f: f is None }})
     r"""Unique ID of company in Codat"""
     id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id'), 'exclude': lambda f: f is None }})