+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/shared/syncinitiated.py
@@ -11,7 +11,6 @@ from typing import Optional
 
 @dataclasses.dataclass
 class SyncInitiated:
-    r"""Returns the newly created SyncId"""
     sync_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('syncId'), 'exclude': lambda f: f is None }})